Remove dot output
Since pydot is no longer compatible with the latest networkx, and there
doesn't seem to have been a lot of use of it, the support for dot output
in twarc-network is being removed. We can revisit this if needed, but
hopefully the other formats will be sufficient.

The recommendation to switch to pygraphviz would bring complications
when installing twarc-network since it requires a C/C++ compiler to build
against the Graphviz library.

closes #13